+++ System/Reactive/Observer.lua
@@ -15,10 +15,6 @@
 PLoop(function(_ENV)
     namespace "System.Reactive"
 
-    --- The default subscription can be created based on root subscription
-    __Sealed__()
-    class "Subscription"                 { System.ISubscription }
-
     --- The default observer
     __Sealed__()
     class "Observer"                    (function(_ENV)
@@ -27,60 +23,43 @@ PLoop(function(_ENV)
         export                          {
             rawset                      = rawset,
             rawget                      = rawget,
-            fakefunc                    = Toolset.fakefunc,
 
             Subscription
-        }
-
-        -----------------------------------------------------------------------
-        --                             property                              --
-        -----------------------------------------------------------------------
-        -- The subscription will be used by the observer
-        property "Subscription"         {
-            type                        = ISubscription,
-            field                       = "__subscription",
-            default                     = function(self)
-                local sub               = Subscription()
-                -- make sure a new subscription will be created if the observer is reused
-                sub.OnUnsubscribe       = sub.OnUnsubscribe + function() rawset(self, "__subscription", nil) end
-                return sub
-            end
         }
 
         -----------------------------------------------------------------------
         --                              method                               --
         -----------------------------------------------------------------------
         --- Provides the observer with new data
-        function OnNext(self, ...)      return self.__onNext(...) end
+        function OnNext(self, ...)
+            local on                    = self.__onNext
+            return on and on(...)
+        end
 
         --- Notifies the observer that the provider has experienced an error condition
-        function OnError(self, ...)     return self.__onError(...)end
+        function OnError(self, ...)
+            local on                    = self.__onError
+            return on and on(...)
+        end
 
         --- Notifies the observer that the provider has finished sending push-based notifications
         function OnCompleted(self)
-            local subscription          = rawget(self, "__subscription")
-            return subscription and subscription:Dispose() or self.__onComp()
+            self.Subscription           = nil
+            local on                    = self.__onComp
+            return on and on()
         end
 
         -----------------------------------------------------------------------
         --                            constructor                            --
         -----------------------------------------------------------------------
-        __Arguments__{ Callable/nil, Callable/nil, Callable/nil, ISubscription/nil }
+        __Arguments__{ Callable/nil, Callable/nil, Callable/nil, Subscription/nil }
         function __ctor(self, onNext, onError, onCompleted, subscription)
-            rawset(self, "__onNext",    onNext      or fakefunc)
-            rawset(self, "__onError",   onError     or fakefunc)
-            rawset(self, "__onComp",    onCompleted or fakefunc)
+            rawset(self, "__onNext",    onNext      or false)
+            rawset(self, "__onError",   onError     or false)
+            rawset(self, "__onComp",    onCompleted or false)
             if subscription then
                 self.Subscription       = Subscription(subscription)
             end
-        end
-
-        -----------------------------------------------------------------------
-        --                          de-constructor                           --
-        -----------------------------------------------------------------------
-        function __dtor(self)
-            local subscription          = rawget(self, "__subscription")
-            return subscription and subscription:Dispose()
         end
     end)
 
@@ -94,25 +73,24 @@ PLoop(function(_ENV)
 
             -- the core subscribe
             subscribe                   = function (self, observer, subscription)
-                subscription            = subscription or isObjectType(observer, Observer) and observer.Subscription or Subscription()
+                subscription            = subscription or observer.Subscription
                 self.__subscribe(observer, subscription)
                 return subscription, observer
             end,
 
-            Observer, Subscription
+            Observer
         }
 
         -----------------------------------------------------------------------
         --                              method                               --
         -----------------------------------------------------------------------
         --- Subscribe the observer with subscription
-        __Arguments__{ IObserver, ISubscription/nil }
+        __Arguments__{ IObserver, Subscription/nil }
         Subscribe                       = subscribe
 
-        __Arguments__{ Callable/nil, Callable/nil, Callable/nil, ISubscription/nil }
+        __Arguments__{ Callable/nil, Callable/nil, Callable/nil, Subscription/nil }
         function Subscribe(self, onNext, onError, onCompleted, subscription)
-            local observer              = Observer(onNext, onError, onCompleted, subscription)
-            return subscribe(self, observer, observer.Subscription)
+            return subscribe(self, Observer(onNext, onError, onCompleted, subscription))
         end
 
         -----------------------------------------------------------------------